Back off and retry whenever API rate limits are reached
Atlassian applies rate limiting to Jira's REST APIs [1]. When the rate
limit for a particular endpoint has been reached, Jira produces a
response that is compatible with RFC 6585 section 4: the status is set
to `429 Too Many Requests`, and the client is expected not to reattempt
the request until at least the number of seconds given by the
`Retry-After` header have elapsed. Currently the default `Client` and
the built-in `Transport`s do not respect any rate limits, and exceeding
them results in errors.

Use `github.com/hashicorp/go-retryablehttp` to implement RFC 6585
section 4 in the default `Client` and built-in `Transport`s so that the
API rate limits are respected. A welcome side-effect is that failed
requests due to (e.g.) network problems will also be retried (up to a
maximum of 4 times, go-retryablehttp's default).

[1] https://developer.atlassian.com/cloud/jira/platform/rate-limiting/