Allow API endpoints to determine default boolean option values
On Jira API endpoints that accept boolean parameters, the value of the
parameter may default to either true or false if it isn't supplied in
the request, depending on the endpoint. In options structs that have
them, change the type of fields representing these parameters from
`bool` to `*bool`, so that their value isn't explicitly set to false in
the request if a value isn't given in the struct - in order words, let
the API endpoint decide what the value should be, not go-jira.